[refactor] validator -> parser로 클래스 변경
