+++ src/main/java/step2/Domain/Rank.java
@@ -1,10 +1,7 @@
 package step2.Domain;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class Rank {
 
@@ -25,13 +22,4 @@ public class Rank {
         return rank;
     }
 
-    public Rank merge(Rank anotherRank){
-        return new Rank(Stream.of(this.rank, anotherRank.getRank())
-                .flatMap(mergedMap -> mergedMap.entrySet().stream())
-                .collect(Collectors.toMap(
-                        Map.Entry::getKey,
-                        v -> new ArrayList<>(v.getValue()),
-                        (left, right) -> {left.addAll(right); return left;}
-                )));
-    }
 }
